fix filtre + test facture pas fini

+++ backendCompteurApp/src/main/java/com/compteurapp/backendcompteurapp/DTO/FactureDTO.java
@@ -3,7 +3,7 @@ package com.compteurapp.backendcompteurapp.DTO;
 import com.compteurapp.backendcompteurapp.enums.FactureStatement;
 
 public class FactureDTO {
-    public int prix;
+    public double prix;
     public FactureStatement etat;
     public Long idCompteurData;
 }
